+++ HISTORY.rst
@@ -5,6 +5,7 @@ History
 0.32.0 (2024-XX-XX)
 -------------------
 
+* Add Leiden clustering algorithm
 * Add functions to detect and break cycles
 * Add damping factor in diffusion
 * Fix F1 scores